delete login service and move to API
