+++ data/src/main/java/com/team_ia/data/remote/api/GetGoogleTokenAPI.kt
@@ -1,4 +1,4 @@
-package com.team_ia.data.service
+package com.team_ia.data.remote.api
 
 import com.team_ia.data.remote.request.auth.LoginGoogleRequest
 import com.team_ia.data.remote.response.auth.LoginGoogleResponse
@@ -6,8 +6,7 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface LoginService {
-
+interface GetGoogleTokenAPI {
     @POST("oauth2/v4/token")
     suspend fun fetchGoogleAuthInfo(
         @Body request: LoginGoogleRequest
